Lena Bojc/review-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.1